HSMM with poisson instead of negative binomial + extra
When I apply the HSMM to my data I often get negative binomial size
estimates that are extremely large, which means that we are approaching
a poisson distribution and is causing tons of numerical problems. I’m
thus creating an option using the poisson distribution which has less
parameters than the negative binomial. Also I’ve made a general
function for the CI and changes some starting values.